app/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
nltk==3.9
transformers==4.33.0
bert-score==0.3.13
python-Levenshtein==0.21.0
numpy==1.24.0
